优化代码与文档
