+++ .gitignore
@@ -8,6 +8,8 @@ __pycache__/
 # C extensions
 *.so
 
+.DS_Store
+
 # node
 /node_modules
