+++ bin/stop.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# 停止 Docker Compose
+docker-compose down
