+++ conf/mysql.cnf
@@ -1,0 +1,3 @@
+[mysqld]
+skip-host-cache
+skip-name-resolve
