+++ conf/nginx.conf
@@ -1,0 +1,36 @@
+worker_processes 1;
+
+events {
+    worker_connections 1024;
+}
+
+http {
+    include mime.types;
+    default_type application/octet-stream;
+    sendfile on;
+
+    server {
+        listen 80;
+        server_name localhost;
+
+        access_log /var/log/nginx/access.log;
+        error_log /var/log/nginx/error.log;
+
+        root /data/home_front;
+        index index.html;
+
+        location / {
+            if ($request_uri ~ ^/(.*)\.html(.*)$) {
+                return 302 /$1$2;
+            }
+            try_files $uri $uri.html $uri/ =404;
+        }
+
+        location /api {
+            proxy_pass http://127.0.0.1:8080;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        }
+    }
+}
