+++ conf/redis.conf
@@ -1,0 +1,3 @@
+bind 0.0.0.0
+protected-mode no
+appendonly yes
