+++ house_rental/dao/__init__.py
@@ -1,0 +1,4 @@
+from .mysql import init_mysql
+from .redis import init_redis
+from .logging import init_logging
+
